+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # resourcet-effectful-1.0.1.0 (2023-??-??)
-* Add `allocateEff`, `allocateEff_` and `registerEff`.
+* Add `allocateEff`, `allocateEff_`, `registerEff`, `releaseEff` and
+  `unprotectEff`.
+* Re-export `ReleaseKey` and `ResourceCleanupException`.
 
 # resourcet-effectful-1.0.0.0 (2022-07-14)
 * Initial release.